pcre32: Fix signed-unsigned compare
git-svn-id: svn://vcs.exim.org/pcre/code/trunk@1134 2f5784b3-3f2a-0410-8824-cb99058d5e15

+++ pcre_fullinfo.c
@@ -136,7 +136,7 @@ switch (what)
 
   case PCRE_INFO_FIRSTBYTE:
   *((int *)where) =
-    ((re->flags & PCRE_FIRSTSET) != 0)? re->first_char :
+    ((re->flags & PCRE_FIRSTSET) != 0)? (int)re->first_char :
     ((re->flags & PCRE_STARTLINE) != 0)? -1 : -2;
   break;
 
@@ -174,7 +174,7 @@ switch (what)
 
   case PCRE_INFO_LASTLITERAL:
   *((int *)where) =
-    ((re->flags & PCRE_REQCHSET) != 0)? re->req_char : -1;
+    ((re->flags & PCRE_REQCHSET) != 0)? (int)re->req_char : -1;
   break;
 
   case PCRE_INFO_REQUIREDCHAR:
